# web/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Koritsu API",
  "description": "Two entrances, one set of handlers: `/api` for the site (cookie session) and `/api/v1` for external clients (bearer token). Only `/api/v1` is versioned and documented.",
  "version": "2.0.0a5.4"
 },
 "paths": {
...
  "/api/projects": {
   "post": {
    "tags": [
     "projects"
    ],
    "summary": "Create a project in a workspace",
    "description": "Creates a project and its directory on the volume. Multipart: the name and, optionally, either a DOCX template or the id of one of your saved templates (`template_id`); without either the report is built from scratch. `module` says which module the work is done with (from GET /api/modules) and may be left out. Editor role in the workspace. 400 bad_template, 400 unknown_module, 403 forbidden, 404 not_found, 409 project_exists.",
    "operationId": "create_project",
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_create_project"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Create Project"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "projects"
    ],
    "summary": "List projects of a workspace",
    "description": "Lists the projects of one workspace; `trash=true` lists the ones in the trash instead, `module` narrows the list down to the works done with one module. 400 invalid_id, 400 unknown_module, 404 not_found.",
    "operationId": "list_projects",
    "parameters": [
     {
      "name": "workspace_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Workspace Id"
      }
     },
     {
      "name": "trash",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Trash"
      }
     },
     {
      "name": "module",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "",
       "title": "Module"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Projects"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/jobs": {
   "post": {
    "tags": [
     "jobs"
    ],
    "summary": "Put a job into the queue",
    "description": "Accepts a job and answers at once with its card; the work happens in a worker process. Editor role in the project. A second job of the same user waits in the queue instead of being refused. 400 unknown_job_kind, 400 project_required, 402 limit_exhausted, 403 forbidden, 404 not_found, 409 in_trash.",
    "operationId": "jobs_create",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/JobIn"
       }
      }
     }
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Jobs Create"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "jobs"
    ],
    "summary": "List your jobs",
    "description": "Your own jobs, newest first. Jobs of other people are never listed, not even in a shared project. 400 invalid_id, 400 unknown_job_status.",
    "operationId": "jobs_list",
    "parameters": [
     {
      "name": "project_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Only jobs of this project",
       "title": "Project Id"
      },
      "description": "Only jobs of this project"
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Only jobs in this state: queued, running, done, failed, cancelled",
       "title": "Status"
      },
      "description": "Only jobs in this state: queued, running, done, failed, cancelled"
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Jobs List"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/tokens": {
   "get": {
    "tags": [
     "tokens"
    ],
    "summary": "List your API tokens",
    "description": "Your tokens, newest first: prefix, name, scopes and dates. Revoked ones stay in the list, marked with revoked_at. The token string is never listed. 401 token_not_allowed.",
    "operationId": "list_api_tokens",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ApiTokenOut"
         },
         "type": "array",
         "title": "Response List Api Tokens"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "tokens"
    ],
    "summary": "Create an API token",
    "description": "Issues a token for the /api/v1 entrance and returns its string once: the service keeps only a hash, so a lost token is revoked and reissued, never recovered. 400 unknown_scope, 401 token_not_allowed.",
    "operationId": "create_api_token",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ApiTokenIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiTokenCreated"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/bootstrap": {
   "get": {
    "tags": [
     "bootstrap"
    ],
    "summary": "Everything the first screen needs, in one answer",
    "description": "The signed-in account, the modules the interface may show, the monthly limit with what is left, the workspaces the caller belongs to, the newest notifications with the unread count, and admin_domain: the host the admin area lives on, or an empty string when it is not split off a subdomain. Every nested object has exactly the shape its own route returns, and is built by the very same code; the separate routes stay for refreshing one thing at a time. 401 unauthenticated.",
    "operationId": "bootstrap",
    "parameters": [
     {
      "name": "notifications",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "description": "how many newest notifications to include",
       "default": 20,
       "title": "Notifications"
      },
      "description": "how many newest notifications to include"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Bootstrap"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/projects": {
   "post": {
    "tags": [
     "v1",
     "projects"
    ],
    "summary": "Create a project in a workspace",
    "description": "Creates a project and its directory on the volume. Multipart: the name and, optionally, either a DOCX template or the id of one of your saved templates (`template_id`); without either the report is built from scratch. `module` says which module the work is done with (from GET /api/modules) and may be left out. Editor role in the workspace. 400 bad_template, 400 unknown_module, 403 forbidden, 404 not_found, 409 project_exists.",
    "operationId": "create_project_v1",
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_create_project_v1"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Create Project V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "v1",
     "projects"
    ],
    "summary": "List projects of a workspace",
    "description": "Lists the projects of one workspace; `trash=true` lists the ones in the trash instead, `module` narrows the list down to the works done with one module. 400 invalid_id, 400 unknown_module, 404 not_found.",
    "operationId": "list_projects_v1",
    "parameters": [
     {
      "name": "workspace_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Workspace Id"
      }
     },
     {
      "name": "trash",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Trash"
      }
     },
     {
      "name": "module",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "",
       "title": "Module"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Projects V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/jobs": {
   "post": {
    "tags": [
     "v1",
     "jobs"
    ],
    "summary": "Put a job into the queue",
    "description": "Accepts a job and answers at once with its card; the work happens in a worker process. Editor role in the project. A second job of the same user waits in the queue instead of being refused. 400 unknown_job_kind, 400 project_required, 402 limit_exhausted, 403 forbidden, 404 not_found, 409 in_trash.",
    "operationId": "jobs_create_v1",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/JobIn"
       }
      }
     }
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Jobs Create V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "v1",
     "jobs"
    ],
    "summary": "List your jobs",
    "description": "Your own jobs, newest first. Jobs of other people are never listed, not even in a shared project. 400 invalid_id, 400 unknown_job_status.",
    "operationId": "jobs_list_v1",
    "parameters": [
     {
      "name": "project_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Only jobs of this project",
       "title": "Project Id"
      },
      "description": "Only jobs of this project"
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Only jobs in this state: queued, running, done, failed, cancelled",
       "title": "Status"
      },
      "description": "Only jobs in this state: queued, running, done, failed, cancelled"
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Jobs List V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "ApiTokenCreated": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id",
      "description": "Token id, used to revoke it"
     },
     "name": {
      "type": "string",
      "title": "Name",
      "description": "Human-readable name"
     },
     "prefix": {
      "type": "string",
      "title": "Prefix",
      "description": "First characters, to tell tokens apart"
     },
     "scopes": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Scopes",
      "description": "What this token may do"
     },
     "created_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Created At",
      "description": "When it was issued"
     },
     "last_used_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Used At",
      "description": "Last use, written at most once a minute"
     },
     "revoked_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Revoked At",
      "description": "When it was revoked, if it was"
     },
     "token": {
      "type": "string",
      "title": "Token",
      "description": "The token string; shown once and never stored in clear"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "prefix",
     "scopes",
     "token"
    ],
    "title": "ApiTokenCreated",
    "description": "Ответ на создание: карточка и сама строка ключа, в первый и последний раз."
   },
   "ApiTokenIn": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 64,
      "title": "Name",
      "description": "Human-readable name, e.g. 'CI' or 'laptop'"
     },
     "scopes": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Scopes",
      "description": "What the token may do; at least one of: projects:read, projects:write, materials:write, runs:run"
     }
    },
    "type": "object",
    "required": [
     "name",
     "scopes"
    ],
    "title": "ApiTokenIn",
    "description": "Тело `POST /api/tokens`: как назвать и что этим ключом можно."
   },
   "ApiTokenOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id",
      "description": "Token id, used to revoke it"
     },
     "name": {
      "type": "string",
      "title": "Name",
      "description": "Human-readable name"
     },
     "prefix": {
      "type": "string",
      "title": "Prefix",
      "description": "First characters, to tell tokens apart"
     },
     "scopes": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Scopes",
      "description": "What this token may do"
     },
     "created_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Created At",
      "description": "When it was issued"
     },
     "last_used_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Used At",
      "description": "Last use, written at most once a minute"
     },
     "revoked_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Revoked At",
      "description": "When it was revoked, if it was"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "prefix",
     "scopes"
    ],
    "title": "ApiTokenOut",
    "description": "Карточка ключа. Строки ключа в ней нет и быть не может."
   },
   "Body_create_project": {
    "properties": {
     "workspace_id": {
      "type": "string",
      "title": "Workspace Id"
     },
     "name": {
      "type": "string",
      "title": "Name",
      "default": ""
     },
     "module": {
      "type": "string",
      "title": "Module",
      "default": ""
     },
     "template": {
      "anyOf": [
       {
        "type": "string",
        "contentMediaType": "application/octet-stream"
       },
       {
        "type": "null"
       }
      ],
      "title": "Template"
     },
     "template_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Template Id"
     }
    },
    "type": "object",
    "required": [
     "workspace_id"
    ],
    "title": "Body_create_project"
   },
...
   "ErrorOut": {
    "properties": {
     "error": {
      "$ref": "#/components/schemas/ErrorBody"
     }
    },
    "type": "object",
    "required": [
     "error"
    ],
    "title": "ErrorOut",
    "description": "Единственная форма отказа службы. Другой нет ни у одного маршрута."
   },
...
   "JobIn": {
    "properties": {
     "kind": {
      "type": "string",
      "title": "Kind",
      "description": "What to do; one of: fill_tag, fill_report, agent, build, parse, export, kadai_run, kadai_rework, probe"
     },
     "project_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Project Id",
      "description": "Project the job works in; required by most kinds"
     },
     "payload": {
      "additionalProperties": true,
      "type": "object",
      "title": "Payload",
      "description": "Arguments of the job; the shape depends on the kind"
     }
    },
    "type": "object",
    "required": [
     "kind"
    ],
    "title": "JobIn",
    "description": "Тело постановки задания.\n\nИмя по-английски, как у соседей (`ProjectPatchIn`, `ModelKeyIn`): оно уезжает\nв OpenAPI и становится именем типа в клиенте сайта."
   },
...
  }
 },
...
}